CS11 - ReyhanBatara

+++ frontend_ReyhanBatara/core/src/main/java/com/Reyhan/frontend/strategies/DifficultyStrategy.java
@@ -4,7 +4,7 @@ import java.util.Map;
 
 public interface DifficultyStrategy {
     float getSpawnInterval();
-    float getDensity();
+    int getDensity();
     float getMinGap();
     Map<String, Integer> getObstacleWeights();
 }
